An attempt at finding screenX when there isn't a right angle triangle between the eye and fixation point.
I'm not sure if the logic of what I've done is sound in the
calculateScreenLocation.m file, but if I test it with specific values it
does produce the result I'm expecting. However, it doesn't work if A is
closer to the fixation point in the X axis than the eye is -- But I'm
not sure if this is a problem or not. Quite a lot to think about.