Coordinate system relative to cyclopean point & object position in two eyes
Moved the (0,0,0) point from the centre of one eye to the cyclopean
point (1/2 of the IOD away in X) and did a test to make sure it worked.
Then added the code for the second eye on the other side of the
cyclopean point and did some tests with it.