Multiple changes to MoveLineTrial and associated files
Added code so that a looming-only stimulus consisting of two horizontal
lines (as discussed in the project meeting back in October) could be
used if we want to rather than covering one eye for the combined
stimulus. I've also created a calculateHorizontalScreenLocation function
that calculates positions for everything in terms of y and z rather than
x and z, and a small piece of code to test that.  There are then some
additions to the paradigm file to allow this to work. I've then tried to
comment everything that is going on in the MoveLineTrial file to make it
clearer what is going on. So there's now three different types of
stimulus that you can choose from within these files.