Changes so that fixation cross is always present.
I've also added a box around the fixation cross when there aren't any
moving lines -- However I haven't been able to separate the "give a
response" time from just the inter-trial interval. I've tried using
exist, isstruct and isfield on trialData.firstCond (and some of the
fields within firstCond) so that the box is only present if that nested
struct exists but I couldn't get it to work.