Changes to psychMaster and paradigm files to allow velocities higher than the null.
Added conditionInfo.velocityLessThanNull to all paradigm files and added
three conditions for null velocity +5%, +10% and +20%.

In psychMaster then used this in combination with nullFirst to work out
which answer would be correct in each condition scenario.

The velocities greater than the null velocity cause the lines to reach
the edge and then reverse direction -- may need to stop the lines from
going off the screen. May also want to make the null less than the
viewing distance.