An analysis file for the MoveLineTrial data
Really basic right now, but will plot a graph of percentage correct
against the condition velocities (as a percentage of the null for now).
Need to edit it so that it can cope with invalid trials, and so that it
automatically reads in data files rather than manually. Will probably
put things into loops.