Changes to analysis and small edits on all paradigm files
Changed the analysis file so that it can cope with invalid trials being
in experimentData and excludes the invalid trials from the percentage
correct responses calculation. I've introduced some code so that it also
produces a second type of graph of the percentage correct responses
against the percentage difference in velocity from the null (5, 10, 20).

In all of the paradigm files I've changed the null distance travelled to
be less than the viewing distance to 40cm so that the null velocity is
8cm/s and that the distance travelled is never more than the viewing
distance -- this has been causing problems.