Changes to output how the subject responded but not correctness
Removed “correct” and “incorrect” response options from the 2afc case
of psychMaster so that you only get what they respond and not whether
or not it was correct. I did this so that we could have the 7th
condition which is equal to the null, which wasn’t possible with
correct/incorrect. I’ve also added the 7th condition to all three of
the paradigm files and re-arranged their order so 0.8 of null is (1),
0.9 of null is (2) equal to null is (4) etc.