Fixation boxes only appear when a participant can make a response
Edited psychMaster and MoveLineTrial so that the boxes would only be
drawn around the fixation cross when the participant can make a
response. Created a new function, drawFixationInfo, which contains the
information needed for drawing the fixation cross and puts it into
expInfo so that it can be accessed in psychMaster and MoveLineTrial
without re-writing it all. Also converted screenXcentre and
screenYcentre into expInfo.center(1) and expInfo.center(2) respectively
as it's the same information being used and it was a bit messy.