Changes to instructions, comments added and quadrupled the trial length
Altered the instructions for the experiment to find which line changed
speed. Added a few comments to the trial file. I couldn't see the speed
change laterally so I made the trial a lot longer by making each section
0.5s (so the whole stim duration is 1s) at which point, without the
stereoscope you can see a speed change for the larger differences.