Temporary changes to MoveLineTrial to allow physical measurement of line movement
A bit of a weird edit to the move line trial file. Wanted to directly
measure if the looming and combined have the same size change on the
screen in the lab. To do this I've added lines to those sections that
draw in black the lines in the line starting position. I've then added
KbStrokeWait at the end of section 2 of movement for both looming and
combined so that the end position is also frozen onto the screen. When
I've checked this in the lab I will comment out these sections so that
normal functioning is back but we can still check this again in the
future.