Created drawFixation; implemented in psychMaster
Created the function drawFixation. The inputs are expInfo, fixationType
(so in the code section you can specify what type of fixation cross you
would like), responseSquare (so if it ==1, you get the square drawn
around the fixation square) and what type of aperture you would like to
draw around the stimulus to aid fixation.
- The aperture section doesn’t do anything yet.
- I haven’t yet figured out how to get this to work for MoveLineTrial —
for now that is still just using drawFixationInfo and drawing the
fixation cross each time it is needed. When I tried to implement it,
the lines for both eyes were drawn on the right hand side only, and
everything flickered as if the flips were messed up.