Randomising the seed before deciding if the null should be displayed first
drawFixation addition had meant that the seed remained fixed for
picking nullFirst and so the condition and null were being displayed in
the same order in every trial; not good. Fixed now with rng(‘shuffle’).
Put it here so that it wouldn’t reintroduce lag problems.

Also changed direction paradigm instructions slightly.