Initial version of a paradigm file for a constant retinal speed task
Added a new paradigm file for this and also edited move line trial
(because the depth thing that affects whether it moves in z or x is
there). Added a new section so that I wouldn't try to do the combined
task not realising I'd forgotten to change it back to moving in z.