Changes to how drawFixation is called
Rather than drawFixation having inputs of expInfo, fixationType,
responseSquare and apetureType, it now has inputs of expInfo and
fixationInfo. FixationInfo is a struct that contains fixationType,
responseSquare and apertureType. Edited everything in drawFixation,
MoveLineTrial and psychMaster that uses the drawFixation code so that
it runs with this new setup without errors.

Also tidied up some of MoveLineTrial and added a more thorough first
comment to appear when using ‘help’ to try to explain what the
different sections of the code actually do.