Introduction of constant retinal speed in depth
Introduced code to MoveLineTrial to allow experiments where there is
constant retinal speed but the stimulus is moving through depth.

Also added a new paradigm file to run and test this which will be
modified to ultimately be the slow-speed version of this experiment.

Also modified the lateral constant retinal speed paradigm files
slightly because I changed the name of the the CRS conditions in
MoveLineTrial.