Initial commit for a sine wave trial file and a corresponding paradigm file
Initial commit for a trial file that has a drifting sine wave grating.
Currently it does not work and contains nothing that would actually draw
a sine wave grating. Trying to get a test piece of code in testSandbox
to draw a static sine wave grating before coming back to these pieces of
code, adapting them further, getting a static grating and then
eventually a drifting grating to work.