Implementation of interval beeps
1 beep = interval 1; 2 beeps = interval 2.

Currently quite clunky, and some sections that may not need to be
repeated are.

But does produce beeps indicating the intervals, and then, if selected
audio feedback. I’ve added a pause time after the end of the audio
feedback, because otherwise the feedback sound and the interval 1 sound
for the next trial are very close together and that may be confusing.

May make the interval sounds optional via paradigm files as well as the
feedback sounds.