Made interval beeps optional
Can choose in the paradigm files whether or not to have the interval
beeps.

Audio feedback will still be given separately if wanted.

If whether or not you want the audio beeps for the intervals and
feedback is not defined in the paradigm file, psychMaster defaults to
not giving audio interval beeps or feedback (depending on which one is
not defined).