Setting the start position in depth in CRS depth paradigm files
manually setting the start position of each line in both constant
retinal speed in depth paradigm files so that they appear to move from
behind the fixation cross towards the observer, rather than from the
plane of fixation.

slow appears to begin 10cm behind fixation. mid appears to begin 20cm
behind fixation. This is the same as for the realistic motion in depth
experiments.