Additions of alternative speed descriptions
for the accelerating lateral and CRS depth conditions I’ve now added
their lateral and depth speeds in cm/s respectively so that they can be
compared with the opposing condition of the same speed. Also made a
script that would calculate the speed ranges of the accelerating
lateral condition.