Fixes to condition labels, the duration of the fixed distance experiment and the depth start
labels now correspond to speeds rather than the speed change of the old
conditions, duration of fixed distance now = 500ms rather than 1s and
I’ve changed the depth start for all of these to 0 so that the arcmin
velocity on the retina is closer to that in the speed change arcmin
velocities. Will need to recalculate this values.