Inclusion of catch trials in speed discrimination paradigms
NEED TO BE TESTED FULLY ON LAB COMPUTER

Adds a level to the fixed distance paradigm that has fixed duration and
adds a level to the fixed duration paradigm that has fixed distance.

Fixed distance has a catch with a slower speed than the null that is
visible for the same number of frames as level 7, so a shorter distance
is travelled over a shorter time. If participants have been responding
to speed, they will get 0% correct in this condition. If they are
responding to duration throughout the block (or the distance travelled
here), they will get 100% correct in this condition.

Fixed duration has a catch where the speed is the same as the null, but
the duration is longer and the distance matches that of the level 7
condition. If people respond at chance level, they are responding to
speed. If people get 0% correct, they are responding to time and if
they get 100% they are responding to distance.