New catch trial setup
Needs checking in the lab.

Balanced catch present in both - speeds the same in test and null, but
one travels a larger distance over a longer duration - 50% correct for
speed responses, 0% correct for duration responses and 100% correct for
distance responses.

Then one other catch trial in each condition.

Fixed distance: slow speed with a short duration catch. If people are
responding to speed, 0% correct; but if they are responding to the
short duration = 100% correct (change strategy to distance responses =
0% correct)

Fixed duration: slow speed with a short null duration catch. If people
responding to speed, 0% correct; but if they are responding to the
longer distance = 100% correct (change strategy to duration responses =
0% correct)