Abigail develop (#14)
* Fixed duration paradigm file

Paradigm file for doing a fixed duration speed discrimination task with
the looming (fast) stimulus

* Fixed distance paradigm file

Paradigm file for fixed duration with edits to MoveLineTrial to get
this to work. Adapts number of frames in each presentation depending on
the velocity so that distance travelled is constant.

Also changed the name of the fixed duration paradigm file.

* Tidying MoveLineTrial

Removing old commented out code that was used for making fixed line
positions but is now confusing

* TrialData includes start and end line position for looming conditions only

the start and end positions are in pixels

* corrected experiment instructions for speed tasks

now ask which one moved faster rather than which one changed speed.

* Fixes to condition labels, the duration of the fixed distance experiment and the depth start

labels now correspond to speeds rather than the speed change of the old
conditions, duration of fixed distance now = 500ms rather than 1s and
I’ve changed the depth start for all of these to 0 so that the arcmin
velocity on the retina is closer to that in the speed change arcmin
velocities. Will need to recalculate this values.

* Speed discrimination training blocks

levels 3 and 7 from speed discrimination blocks to be used as
screening/training for participants. Includes audio feedback. 30 trials
per level.

* Inclusion of catch trials in speed discrimination paradigms

NEED TO BE TESTED FULLY ON LAB COMPUTER

Adds a level to the fixed distance paradigm that has fixed duration and
adds a level to the fixed duration paradigm that has fixed distance.

Fixed distance has a catch with a slower speed than the null that is
visible for the same number of frames as level 7, so a shorter distance
is travelled over a shorter time. If participants have been responding
to speed, they will get 0% correct in this condition. If they are
responding to duration throughout the block (or the distance travelled
here), they will get 100% correct in this condition.

Fixed duration has a catch where the speed is the same as the null, but
the duration is longer and the distance matches that of the level 7
condition. If people respond at chance level, they are responding to
speed. If people get 0% correct, they are responding to time and if
they get 100% they are responding to distance.

* Changes so speed matches null in both catches

so if in either catch trial, people are responding to speed, they will
be at chance level

* New catch trial setup

Needs checking in the lab.

Balanced catch present in both - speeds the same in test and null, but
one travels a larger distance over a longer duration - 50% correct for
speed responses, 0% correct for duration responses and 100% correct for
distance responses.

Then one other catch trial in each condition.

Fixed distance: slow speed with a short duration catch. If people are
responding to speed, 0% correct; but if they are responding to the
short duration = 100% correct (change strategy to distance responses =
0% correct)

Fixed duration: slow speed with a short null duration catch. If people
responding to speed, 0% correct; but if they are responding to the
longer distance = 100% correct (change strategy to duration responses =
0% correct)

* bug fixes and minor edits

* bug fixes

* Edit to condition 9 to make it work the way we want

* Made the screening conditions much shorter

too long as training with 30 reps per level; now 15. Just to get people
used to what they are doing, this will give 60 training things with
feedback. Previously experiments also had 60 trials with feedback.